Patch attr nomodule
